requirements.txt
Flask-Mail>=0.9.1
Flask-Migrate>=2.4.0
Flask-SQLAlchemy>=2.4.0
Flask-Script>=2.0.6
SQLAlchemy-Utils>=0.33.11
celery>=4.2.1
connexion>=2.2.0
connexion-buzz>=0.2.1
dynaconf>=2.0.3
flask-marshmallow>=0.10.1
marshmallow-sqlalchemy>=0.16.2
mysqlclient>=1.3.13
redis>=3.2.0
requests>=2.21.0
sentry-sdk[flask]>=0.7.10
uWSGI>=2.0.18